# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customer/wallet:
    get:
      operationId: getCustomerWallet
      tags: [orders]
      summary: Get the authenticated customer's wallet balance and paginated ledger
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Wallet balance and ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerWallet"
        "400":
          description: Invalid pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/{id}/cancel:
    post:
      operationId: cancelCustomerOrder
      tags: [orders]
      summary: Cancel a customer-owned order before preparation starts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderActionResult"
        "202":
          description: Provider refund outcome requires reconciliation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderActionResult"
        "400":
          description: Invalid order or cancellation is no longer allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Payment provider rejected the refund
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/refund:
    post:
      operationId: createCustomerRefundRequest
      tags: [orders]
      summary: Submit a full-order wallet refund request for admin review
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequestInput"
      responses:
        "201":
          description: Refund request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefundRequest"
        "400":
          description: Order is not eligible or reason is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A refund request already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    OrderStatus:
      type: string
      enum: [pending, confirmed, preparing, ready, picked_up, delivered, cancelled]
...
    PaymentStatus:
      type: string
      enum: [pending, paid, failed, refunded]
...
    OrderActionResult:
      type: object
      required: [id, status, paymentStatus, message]
      properties:
        id:
          type: number
        status:
          $ref: "#/components/schemas/OrderStatus"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        message:
          type: string
    RefundRequestStatus:
      type: string
      enum: [pending, processing, approved, rejected, failed]
    RefundRequestInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 3
          maxLength: 1000
    RefundRequest:
      type: object
      required: [id, orderId, amount, reason, status, createdAt]
      properties:
        id:
          type: number
        orderId:
          type: number
        amount:
          type: number
        reason:
          type: string
        status:
          $ref: "#/components/schemas/RefundRequestStatus"
        createdAt:
          type: string
          format: date-time
    WalletTransaction:
      type: object
      required: [id, type, amount, description, referenceType, referenceId, balanceAfter, createdAt]
      properties:
        id:
          type: number
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        description:
          type: string
        referenceType:
          type: string
          enum: [refund, order_payment, admin_adjustment]
        referenceId:
          type: number
        balanceAfter:
          type: number
        createdAt:
          type: string
          format: date-time
    CustomerWallet:
      type: object
      required: [balance, page, pageSize, total, transactions]
      properties:
        balance:
          type: number
        page:
          type: number
        pageSize:
          type: number
        total:
          type: number
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string